+++ src/Entities/PlayerCheckPoint.h
@@ -27,7 +27,7 @@ class PlayerCheckPoint : public Entity
          if (index == ID("rndr"))
             return 20;
          else
-            return (int)this;
+            return (uintptr_t)this;
       }
 
       Entity*
